requirements/_base.txt
# Core
django==3.*
djangorestframework==3.*
Pillow==7.*
django-memcache-status==2.*

# Needed by the doc views
coreapi==2.*
coreschema==0.0.*
uritemplate==3.*

# database
mysqlclient==2.*

# Cache system
python-memcached>=1.*

# Internationalization
django-rosetta
